[#26] Removed unused files. Moved CreateCustomerService to Services::CreateCustomer
